#63: KakaoMap 공통 프로퍼티 작성 및 MapTown 리팩토링
