+++ Zatch/domain/ViewControllers/Chatting/Appointment/MapMeetingViewController.swift
@@ -7,14 +7,6 @@
 
 import UIKit
 
-class MapMeetingViewController: UIView {
-
-    /*
-    // Only override draw() if you perform custom drawing.
-    // An empty implementation adversely affects performance during animation.
-    override func draw(_ rect: CGRect) {
-        // Drawing code
-    }
-    */
+class MapMeetingViewController: KakaoMapViewController {
 
 }
