+++ Zatch/domain/Views/TownView/MapTownView.swift
@@ -7,35 +7,12 @@
 
 import UIKit
 
-class MapTownView: UIView {
-    
-    let backBtnView = UIView().then{
-        $0.backgroundColor = .white
-        $0.layer.cornerRadius = 40/2
-        $0.layer.shadowColor = UIColor(red: 0, green: 0, blue: 0, alpha: 0.3).cgColor
-        $0.layer.shadowOpacity = 1
-        $0.layer.shadowRadius = 5
-        $0.layer.shadowOffset = CGSize(width: 0, height: 0)
-    }
-    
-    lazy var backBtn = UIButton().then{
-        $0.setImage(UIImage(named: "town_arrow"), for: .normal)
-    }
-    
-    let mapView = MTMapView().then{
-        $0.baseMapType = .standard
-    }
-    
-    lazy var townSettingBtn = Purple52Button(title: "재치 시작하기").then{
-        $0.layer.shadowColor = UIColor(red: 0, green: 0, blue: 0, alpha: 0.4).cgColor
-        $0.layer.shadowOpacity = 1
-        $0.layer.shadowRadius = 8
-        $0.layer.shadowOffset = CGSize(width: 0, height: 2)
-    }
+class MapTownView: KakaoMapView {
 
     override init(frame: CGRect) {
         super.init(frame: .zero)
         
+        initSetting()
         setUpView()
         setUpConstriant()
     }
@@ -43,39 +20,8 @@ class MapTownView: UIView {
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-
-    func setUpView(){
-        
-        self.addSubview(mapView)
-        
-        mapView.addSubview(backBtnView)
-        mapView.addSubview(townSettingBtn)
-        
-        backBtnView.addSubview(backBtn)
-        
-    }
     
-    func setUpConstriant(){
-        
-        backBtnView.snp.makeConstraints{
-            $0.width.height.equalTo(40)
-            $0.top.equalToSuperview().offset(88)
-            $0.leading.equalToSuperview().offset(35)
-        }
-        
-        backBtn.snp.makeConstraints{
-            $0.width.height.equalTo(16)
-            $0.centerX.centerY.equalToSuperview()
-        }
-        
-        mapView.snp.makeConstraints{
-            $0.top.bottom.leading.trailing.equalToSuperview()
-        }
-        
-        townSettingBtn.snp.makeConstraints{
-            $0.bottom.equalToSuperview().offset(-86)
-            $0.leading.equalToSuperview().offset(31)
-            $0.trailing.equalToSuperview().offset(-31)
-        }
+    func initSetting(){
+        self.settingBtn.setTitle("재치 시작하기", for: .normal)
     }
 }
